# api/openapi/techstack-v1.yaml
openapi: 3.1.0
info:
  title: kombify-TechStack API
  description: |
    REST API for kombify-TechStack -- the hybrid cloud control plane for homelabs.

    kombify-TechStack manages infrastructure through a spec-driven architecture:
    validate YAML specs via CUE, generate OpenTofu configurations, and manage
    nodes through a gRPC agent protocol.

    ## Authentication

    Authenticated endpoints accept an OIDC/JWKS bearer token or an API key
    where documented. Self-hosted TechStack defaults to Pocket ID as identity
    provider. SaaS routes enter through the edge with Auth0 as the platform
    identity provider. PocketBase session exchange remains legacy migration
    compatibility only.

    ## Response Format

    **Success:** `{ "data": <object>, "meta": { "total": int, "page": int, "per_page": int } }`

    **Error:** `{ "error": { "code": "<ERROR_CODE>", "message": "...", "details": <any> } }`

    ## Pagination

    List endpoints accept `page` (default: 1) and `per_page` (default: 20, max: 100) query parameters.
  version: 1.0.0
  contact:
    name: kombify
    url: https://kombify.io
  license:
    name: Apache-2.0 AND GPL-3.0
    url: https://github.com/kombifyio/TechStack/blob/main/LICENSE

# ============================================================
# SERVERS
# ============================================================
# All paths in this document are authored as /api/v1/*. The Gateway projects
...
servers:
  - url: https://api.kombify.io/v1/techstack
...

...
security:
  - bearerAuth: []
paths:
...
  /api/v1/agents/{id}:
    get:
      operationId: getAgent
      tags: [Agents]
      summary: Get agent by ID with resources
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeAgent
      tags: [Agents]
      summary: Remove/deregister agent
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Agent removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  agentId:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC/JWKS bearer token. Self-hosted defaults to Pocket ID; SaaS uses Auth0 through the edge. PocketBase JWT is legacy migration compatibility only.
    runtimeAgentAuth:
      type: http
      scheme: bearer
      bearerFormat: TechStack-Runtime-Agent
      description: Runtime-agent credential bound to the exact tenant, placement, server, and worker identity. Signed, opaque, and explicit legacy compatibility tokens are validated fail-closed by the worker runtime.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service accounts and automation clients (SaaS mode via the public edge)
    edgeTrustHeader:
      type: apiKey
      in: header
      name: X-Edge-Internal-Trust
      description: Legacy trust header injected by the public edge during the migration cutover (SaaS mode only)
  parameters:
...
    resourceId:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        data:
          description: Response payload
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INTERNAL_ERROR
                - NOT_FOUND
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - CONFLICT
                - VALIDATION_ERROR
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - METHOD_NOT_ALLOWED
            message:
              type: string
            details:
              description: Additional error context (validation errors, etc.)
    ResponseMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        managed_runtime_inventory_complete:
          type: boolean
          description: True only when the managed RuntimeLease authority was configured and read successfully for this inventory response.
        worker_inventory_sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: Stable digest of the complete worker identity, attachment, freshness, and managed RuntimeLease authority snapshot used for pagination.
        managed_runtime_active_lease_ids:
          type: array
          items:
            type: string
          description: Canonically sorted IDs of visible managed RuntimeLeases classified as active for inventory and cleanup. This visibility projection is not capacity reservation or release authority.
        managed_runtime_inactive_lease_ids:
          type: array
          items:
            type: string
          description: Canonically sorted IDs of visible managed RuntimeLeases classified as cancelled or archived for inventory and cleanup. Inactive state does not prove provider absence or release managed-server capacity.
        managed_runtime_lease_generation_digests:
          type: array
          items:
            type: object
            required: [lease_id, resource_generation_digest]
            properties:
              lease_id:
                type: string
              resource_generation_digest:
                type: string
                pattern: '^[0-9a-f]{64}$'
          description: Canonically sorted opaque bindings from each visible managed RuntimeLease ID to its exact provider-resource generation. Missing coverage makes destructive inventory proof invalid while leaving ordinary dashboard inventory readable.
        managed_runtime_duplicate_lease_ids:
          type: array
          items:
            type: string
          description: Managed RuntimeLease IDs represented by multiple persisted worker rows before UI projection deduplication; must be empty for destructive inventory proof.
        managed_runtime_attachment_conflict_lease_ids:
          type: array
          items:
            type: string
          description: Canonically sorted caller-visible managed RuntimeLease IDs with a conflicting tenant-wide worker attachment; must be empty for destructive inventory proof and never exposes foreign private lease or worker details.
...